factor out login etc. from app.py
