+++ app.py
@@ -20,7 +20,7 @@ app = Flask(__name__)
 app.config["SQLALCHEMY_DATABASE_URI"] = env("SQLALCHEMY_DATABASE_URI")
 app.config["SECRET_KEY"] = env("SECRET_KEY")
 login = LoginManager(app)
-login.login_view = "login"
+login.login_view = "login.login"
 from models import *
 
 db.init_app(app)
@@ -40,46 +40,6 @@ from decorators import *
 def home():
     images = Image.query.order_by(desc(Image.date)).all()
     return render_template("recent.html", images=images)
-
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user is None or not user.check_password(form.password.data):
-            return "Invalid username or password"
-        login_user(user, remember=form.remember.data)
-        return redirect(url_for("home"))
-    return render_template("login.html", form=form)
-
-
-@app.route("/signup", methods=["GET", "POST"])
-def signup():
-    if current_user.is_authenticated:
-        return redirect(url_for("home"))
-    form = SignupForm()
-    if form.validate_on_submit():
-        user = User(
-            username=form.username.data,
-            is_admin=bool(env("IS_ADMIN")),
-            can_view=bool(env("CAN_VIEW")),
-            can_upload=bool(env("CAN_UPLOAD"))
-        )
-        user.set_password(form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        flash("Signup successful.")
-        return redirect(url_for("login"))
-    return render_template("signup.html", form=form)
-
-
-@app.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for("login"))
 
 
 @app.route("/<image_hash>")
@@ -113,8 +73,9 @@ def search():
 blueprint = Blueprint('images', __name__, static_url_path='/images', static_folder='images/')
 app.register_blueprint(blueprint)
 
-import admin, upload, album
+import admin, upload, album, login
 
 app.register_blueprint(admin.blueprint, url_prefix='/admin')
 app.register_blueprint(upload.blueprint, url_prefix='/upload')
 app.register_blueprint(album.blueprint, url_prefix='/album')
+app.register_blueprint(login.blueprint, url_prefix='/')
